# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ons — Matching Platform API
  description: |
    Privacy-oriented couple matching backend for the Ons (أنس) platform.

    **Privacy model:** Instagram handles are never stored in the applicant profile.
    They are AES-256-GCM encrypted in a separate `identities` collection and
    resolvable only by authorized admin roles (every access is audit-logged).
    Applicants are identified throughout the system by a human-friendly alias
    such as "Blue Falcon" or "Silent River".

    **Authentication:**
    - Admin endpoints are protected by a session cookie (`admin_token`) set by
      `POST /api/v1/admin/login`. The cookie is `HttpOnly` and `SameSite=Lax`.
      API clients may alternatively pass the raw JWT in the `Authorization: Bearer` header.
    - Applicant profile endpoints use a separate JWT with `type: "applicant"` in the payload,
      issued by `POST /api/v1/profile/login`. Pass it as `Authorization: Bearer <token>`.
      Admin JWTs are rejected on applicant endpoints and vice-versa.

    **Rate limits:**
    - Public form submission: 100 requests per 10 minutes per IP
    - Admin login: 10 requests per minute per IP
    - Admin endpoints: 200 requests per minute per IP
    - Profile login: 10 requests per minute per IP
    - Profile endpoints: 100 requests per minute per IP
  version: 1.0.0
  contact:
    name: Admin
servers:
  - url: http://localhost
    description: Docker — nginx reverse proxy (default)
...
    description: Production
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: admin_token
      description: |
        HttpOnly session cookie set by `POST /api/v1/admin/login`.
        The browser sends it automatically on every request to `/api/v1/admin/*`.
        The cookie is invisible to JavaScript (`HttpOnly`), `SameSite=Lax`, and
        `Secure` in production. Valid for 8 hours (configurable via `JWT_EXPIRY`).

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Fallback for API clients and tooling that cannot send cookies.
        Pass the raw JWT from `POST /api/v1/admin/login` (when testing via curl or
        this Swagger UI). The server tries the cookie first; the header is used only
        when no cookie is present.
        Claims: `sub` = admin MongoDB `_id`, `role` = one of `super_admin | admin | viewer`.

    ApplicantBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /api/v1/profile/login` for applicant self-service access.
        Pass it as `Authorization: Bearer <token>`.
        Claims: `sub` = applicant MongoDB `_id`, `alias` = human-friendly codename,
        `type` = `"applicant"` (distinguishes from admin JWTs which carry `role` instead).
        Valid for 30 days. Admin JWTs are rejected on all `/api/v1/profile/*` endpoints.

    SubmissionKey:
      type: apiKey
      in: header
      name: X-Submission-Key
      description: |
        64-character hex key obtained from `GET /api/v1/form/questionnaire`.
        Call that endpoint first, copy the `submissionKey` from the response,
        then paste it here before trying out `POST /submit`.
  responses:
    Unauthorized:
      description: Missing or invalid session (cookie absent/expired or bad Bearer token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Unauthorized"

  # ─── Reusable schemas ────────────────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          example: "Internal server error"
    ValidationErrorResponse:
      type: object
      required: [success, error, details]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          example: "Validation failed"
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            age: ["Must be at least 18 years old"]
            disclaimer_agreed: ["You must agree to the disclaimer"]

    # ── Form ──────────────────────────────────────────────────────────────────
...
    QuestionnaireQuestion:
      type: object
      required: [id, label, type, sensitive, required, order]
      properties:
        id:
          type: string
          example: "instagram_handle"
        label:
          type: string
          example: "Instagram Handle"
        type:
          type: string
          enum: [text, number, select, multiselect, range, boolean, textarea]
        sensitive:
          type: boolean
          description: "If true, the answer is encrypted and stored separately."
        required:
          type: boolean
        order:
          type: integer
        options:
          type: array
          items:
            type: string
          description: "Only present for select/multiselect questions."
        min:
          type: number
          description: "Only present for number/range questions."
        max:
          type: number
        placeholder:
          type: string
    QuestionnaireSection:
      type: object
      required: [id, title, order, questions]
      properties:
        id:
          type: string
          example: "identity"
        title:
          type: string
          example: "Your Identity"
        order:
          type: integer
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionnaireQuestion'
...
    CreateQuestionnaireRequest:
      type: object
      required: [version, name, sections]
      properties:
        version:
          type: string
          pattern: '^\d+\.\d+\.\d+$'
          example: "2.0.0"
          description: "Must be a new version — cannot overwrite an existing one."
        name:
          type: string
          example: "Matching Form v2"
        sections:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/QuestionnaireSection'
...
    AdminLoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          format: password
          example: "s3cr3t"
    AdminLoginResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
      description: |
        The JWT is not returned in the body. It is delivered as an `HttpOnly`
        cookie (`admin_token`) via the `Set-Cookie` response header.
    AdminMeResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required: [adminId, adminUsername, adminRole]
          properties:
            adminId:
              type: string
              description: MongoDB `_id` of the authenticated admin.
              example: "507f1f77bcf86cd799439011"
            adminUsername:
              type: string
              description: Username of the authenticated admin.
              example: "alice"
            adminRole:
              type: string
              enum: [super_admin, admin, viewer]
              example: "admin"

    # ── Applicant ─────────────────────────────────────────────────────────────
    ApplicantStatus:
      type: string
      enum: [applied, matched, dating, inactive]
      description: |
        - `applied` — submitted the form; awaiting first matching run
        - `matched` — has ≥1 proposed match; still in matching pool
        - `dating` — actively dating one match; excluded from new runs
        - `inactive` — deactivated (self or admin); deletion scheduled in 180 days
    Applicant:
      type: object
      required: [_id, alias, questionnaireVersion, answers, status, createdAt, updatedAt]
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
        alias:
          type: string
          example: "Blue Falcon"
        questionnaireVersion:
          type: string
          example: "1.0.0"
        answers:
          type: object
          description: "All non-sensitive answers. instagram_handle is never present."
          additionalProperties: true
          example:
            location: "Paris, France"
            age: 27
            height_cm: 178
            work: "Software Engineer"
            relationship_type: "Long Term"
        status:
          $ref: '#/components/schemas/ApplicantStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicantListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Applicant'
        pagination:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
            total:
              type: integer
              example: 142
            totalPages:
              type: integer
              example: 8
    ApplicantResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          $ref: '#/components/schemas/Applicant'
    IdentityResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          required: [alias, instagramHandle, applicantId]
          properties:
            alias:
              type: string
              example: "Blue Falcon"
            instagramHandle:
              type: string
              example: "@johndoe"
              description: "Decrypted Instagram handle. This response is audit-logged."
            applicantId:
              type: string
              example: "64f1a2b3c4d5e6f7a8b9c0d1"

    # ── Audit log ─────────────────────────────────────────────────────────────
    AuditLogEntry:
      type: object
      required: [_id, adminId, action, ipAddress, userAgent, timestamp]
      properties:
        _id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d2"
        adminId:
          type: string
          description: |
            MongoDB `_id` of the actor who performed the action.
            For all admin-initiated actions this is the admin's `_id`.
            For `APPLICANT_REVEAL_IDENTITY` this is the applicant's `_id` —
            check `metadata.actorType` to distinguish the two cases.
          example: "507f1f77bcf86cd799439011"
        action:
          type: string
          enum:
            - RESOLVE_IDENTITY
            - LIST_APPLICANTS
            - VIEW_APPLICANT
            - DEACTIVATE_APPLICANT
            - ADMIN_LOGIN
            - APPLICANT_REVEAL_IDENTITY
          description: |
            `APPLICANT_REVEAL_IDENTITY` is written when an applicant calls
            `POST /api/v1/profile/matches/:id/contact` to reveal their partner's
            Instagram handle. In this case `adminId` holds the applicant's MongoDB `_id`
            and `metadata.actorType` is `"applicant"` — use that field to distinguish
            applicant-initiated entries from admin-initiated ones in audit queries.
        targetAlias:
          type: string
          example: "Blue Falcon"
        targetApplicantId:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
        ipAddress:
          type: string
          example: "203.0.113.42"
        userAgent:
          type: string
          example: "Mozilla/5.0 ..."
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    AuditLogListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        pagination:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    # ── Matching ──────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/admin/login:
    post:
      tags: [Admin – Auth]
      summary: Admin login
      operationId: adminLogin
      description: |
        Authenticates an admin against the `admins` collection (bcrypt-verified).
        On success, sets an `HttpOnly` `admin_token` session cookie — the JWT is
        **not** present in the response body. The login action is recorded in the
        audit log.

        **Rate limit:** 20 requests per minute per IP.

        > **Swagger UI note:** browsers cannot read the `HttpOnly` cookie set here,
        > so Swagger UI cannot automatically send it on subsequent calls. To test
        > protected endpoints in this UI, use the **BearerAuth** scheme: copy the
        > raw token from the `Set-Cookie` header (visible in DevTools) and paste it
        > under the **Authorize** button.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequest'
      responses:
        '200':
          description: Login successful — session cookie set
          headers:
            Set-Cookie:
              description: |
                `admin_token=<jwt>; Path=/api/v1/admin; HttpOnly; SameSite=Lax`
                (`Secure` added in production). Expires after `JWT_EXPIRY` (default 8 h).
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
              example:
                success: true
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Invalid credentials"
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/admin/me:
    get:
      tags: [Admin – Auth]
      summary: Session probe — current admin identity
      operationId: adminMe
      security:
        - CookieAuth: []
        - BearerAuth: []
      description: |
        Returns the authenticated admin's ID and role from the verified JWT claims.
        Used by the frontend on page load to determine whether the session cookie
        is still valid without storing any credential in JavaScript.

        Returns `401` if the cookie is absent or expired.
      responses:
        '200':
          description: Active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminMeResponse'
              example:
                success: true
                data:
                  adminId: "507f1f77bcf86cd799439011"
                  adminUsername: "alice"
                  adminRole: "admin"
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Admin – Applicants ───────────────────────────────────────────────────────
  /api/v1/admin/applicants:
    get:
      tags: [Admin – Applicants]
      summary: List applicants
      operationId: listApplicants
      security:
        - CookieAuth: []
        - BearerAuth: []
      description: |
        Returns a paginated list of applicant profiles. No sensitive data
        (instagram handles) is included. Audit-logged as LIST_APPLICANTS.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ApplicantStatus'
          example: applied
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated applicant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicantListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/applicants/{id}:
    get:
      tags: [Admin – Applicants]
      summary: Get an applicant by ID
      operationId: getApplicant
      security:
        - CookieAuth: []
        - BearerAuth: []
      description: "Returns the full applicant profile (no sensitive data). Audit-logged as VIEW_APPLICANT."
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
      responses:
        '200':
          description: Applicant profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicantResponse'
        '400':
          description: Invalid ObjectId format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Applicant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Admin – Applicants]
      summary: Deactivate an applicant
      operationId: deactivateApplicant
      security:
        - CookieAuth: []
        - BearerAuth: []
      description: |
        Sets the applicant status to `inactive`. Does not delete data.
        Deletion is scheduled 180 days from deactivation.
        Audit-logged as DEACTIVATE_APPLICANT.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
      responses:
        '200':
          description: Applicant deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  message:
                    type: string
                    example: "Applicant deactivated successfully."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Applicant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/admin/applicants/{id}/identity:
    get:
      tags: [Admin – Applicants]
      summary: Resolve applicant identity (decrypt Instagram handle)
      operationId: resolveIdentity
      security:
        - CookieAuth: []
        - BearerAuth: []
      description: |
        Decrypts and returns the real Instagram handle for an applicant.

        🔒 **Requires `super_admin` role.** Regular `admin` and `viewer` roles receive 403.

        ⚠️ **Every call to this endpoint is recorded in the audit log** with the
        admin ID, IP address, user-agent, timestamp, and target alias.
        Use only when necessary for matchmaking coordination.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
      responses:
        '200':
          description: Decrypted identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityResponse'
              example:
                success: true
                data:
                  alias: "Blue Falcon"
                  instagramHandle: "@johndoe"
                  applicantId: "64f1a2b3c4d5e6f7a8b9c0d1"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient role — super_admin required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Forbidden"
        '404':
          description: Identity record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Admin – Questionnaires ───────────────────────────────────────────────────
  /api/v1/admin/questionnaires:
    post:
      tags: [Admin – Questionnaires]
      summary: Create questionnaire
      operationId: createQuestionnaire
      security:
        - CookieAuth: []
        - BearerAuth: []
      description: |
        Creates a new questionnaire and atomically deactivates all existing ones.
        Only one questionnaire can be active at a time — this is enforced server-side.
        The new version must not already exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuestionnaireRequest'
            example:
              version: "2.0.0"
              name: "Matching Form v2"
              sections:
                - id: "identity"
                  title: "Your Identity"
                  order: 1
                  questions:
                    - id: "instagram_handle"
                      label: "Instagram Handle"
                      type: "text"
                      sensitive: true
                      required: true
                      order: 1
                      placeholder: "@yourhandle"
      responses:
        '201':
          description: Questionnaire created and previous versions deactivated
          content:
            application/json:
              schema:
                type: object
                required: [success, id, version, deactivatedCount, message]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  id:
                    type: string
                    example: "64f1a2b3c4d5e6f7a8b9c0d1"
                  version:
                    type: string
                    example: "2.0.0"
                  deactivatedCount:
                    type: integer
                    description: "Number of previously active questionnaires deactivated."
                    example: 1
                  message:
                    type: string
                    example: "Questionnaire v2.0.0 created. 1 previous version(s) deactivated."
        '400':
          description: Version already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: "Questionnaire version 2.0.0 already exists."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  # ── Admin – Audit ─────────────────────────────────────────────────────────────
  /api/v1/admin/audit-logs:
    get:
      tags: [Admin – Audit]
      summary: List audit logs
      operationId: listAuditLogs
      security:
        - CookieAuth: []
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Matching ──────────────────────────────────────────────────────────────────
...